Move ProductSelectorViewController to own file
