+++ WooCommerce/Classes/ViewRelated/Products/ProductSelector/ProductSelectorView.swift
@@ -1,24 +1,6 @@
 import SwiftUI
 import Yosemite
 
-/// Hosting controller for `ProductSelectorView`.
-///
-final class ProductSelectorViewController: UIHostingController<ProductSelectorView> {
-    init(configuration: ProductSelectorView.Configuration,
-         source: ProductSelectorView.Source,
-         viewModel: ProductSelectorViewModel) {
-
-        super.init(rootView: ProductSelectorView(configuration: configuration,
-                                                 source: source,
-                                                 isPresented: .constant(true),
-                                                 viewModel: viewModel))
-    }
-
-    @available(*, unavailable)
-    required dynamic init?(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-}
 /// View showing a list of products to select.
 ///
 struct ProductSelectorView: View {
